Iškeltas particularItem algoritmas iš kontrolerio
